Add support for global variable binding in Python.
WARNING: Python does not support getter/setter at module level so when modifying a global variable the C object is not affected. This will need to be solved in probably the same way as SWIG does it by having a global object responsible R/W access to C global variables.

Closes #4